Refactor AddApplicationForm component
