+++ pkg/app/web/src/components/add-application-drawer.stories.tsx
@@ -1,16 +1,17 @@
 import React from "react";
-import { AddApplicationForm } from "./add-application-form";
+import { AddApplicationDrawer } from "./add-application-drawer";
 import { action } from "@storybook/addon-actions";
 import { createDecoratorRedux } from "../../.storybook/redux-decorator";
 
 export default {
-  title: "APPLICATION/AddApplicationForm",
-  component: AddApplicationForm,
+  title: "APPLICATION/AddApplicationDrawer",
+  component: AddApplicationDrawer,
   decorators: [createDecoratorRedux({})],
 };
 
 export const overview: React.FC = () => (
-  <AddApplicationForm
+  <AddApplicationDrawer
+    open
     projectName="pipe-cd"
     onSubmit={action("onSubmit")}
     onClose={action("onClose")}
